requirements.txt
pyyaml>=6.0
requests>=2.28.0
tavily-python>=0.3.3
